Fix: 출력 메세지 조정을 위한 버그 수정  - 상수 순서 수정  - 메세지 버그 수정

+++ src/main/java/lotto/auto/MatchNumberMap.java
@@ -23,8 +23,8 @@ public class MatchNumberMap {
         return matchNumberMap.get(rank);
     }
 
-    public int profit() {
-        int profit = INIT_NUM;
+    public float profit() {
+        float profit = INIT_NUM;
         for (Rank keyRank : matchNumberMap.keySet()) {
             profit += keyRank.getWinningMoney() * matchNumberMap.get(keyRank);
         }
